+++ CHANGELOG.md
@@ -1,13 +1,44 @@
 # CHANGELOG
 
-## 1.1.1
+## 2.0.0
+
+### Releases
+
+- 🚀 RELEASED: Version 2.0.0!
+
+## 1.2.0
 
 ### Features
 
-- 👾 FIXED: Sponsors links
+- ⬆ UPDATED: Optimized Navbar for both Coach & Pro Screens 
+- ⬆ UPDATED: Optimized Banner for both Coach & Pro Screens 
+- ⬆ UPDATED: Styling for Best Results Component
+- ⬆ UPDATED: Table responsiveness in Results table
+- ⬆ UPDATED: Optimized Contact for both Coach & Pro Screens 
+- ⬆ UPDATED: Optimized Footer for both Coach & Pro Screens 
+- ✅ ADDED: Coach Screen and its components
+- ⬆ UPDATED: Optimized LegalWarningScreen for both Cookies, Privacy Policy & Legal Warnings 
+
+## 1.1.2
+
+### Fixes
+
+- ⬆ UPDATED: Project migrated to React
 
 ## 1.1.1
+
+### Fixes
+
+- 👾 FIXED: Sponsors links
+
+## 1.1.0
 
 ### Features
 
 - ⬆ UPDATED: 2024 Calendar and sponsors
+
+## 1.0.0
+
+### Features
+
+- 🚀 RELEASED: First Version!